Fix panic when showing a diff which contains a diff
When committing e.g. a patch-file in a git repository that has a line
which contains `--- ` only, `delta` fails with the following error (as it
expects such a line to contain keywords for the diff):

```
thread 'delta::tests::test_diff_in_diff' panicked at 'byte index 2 is out of bounds of ``', src/libcore/str/mod.rs:2017:9
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace.
```

This patch changes the diffing behavior to ensure that lines containing
`--- ` only won't be parsed.