Hack: add boolean to Style struct to identify non-emph styles
This isn't a correct approach. The problem being addressed is that
currently, Style instances are being used to represent things like the
presence of an emph section in a line with edit annotations. A
consequence is that if an emph and non-emph section are styled the
same, then it's impossible to tell that the sections are different in
any way at all. I think the correct approach is probably to start
using a Struct that models elements such as emph/non-emph to represent
these things, and make each such struct instance possess a style.