Option to set the background extension mode to ANSI or spaces
In side-by-side mode, if `background_color_extends_to_terminal_width`
is set, the left panel color is extended via spaces, but the right
one via an ANSI sequence which instructs the terminal emulator to
fill the background color rightwards.

The command line option --line-fill-method ansi|spaces can change
how the right panel background is filled.

Add enums `BgShouldFill` and `BgFillMethod` to better distinguish
if the background should be filled, and if so, how.