Use file-modified-label consistently.
When filtering a diff and when directly diffing files/dirs, use the
modified label that is output when interpreting git diffs.  This gets
rid of the "comparing:" prefix (which was otherwise missing from the
--navigate regex) and ensures that a direct diff has a label to stop at
when --navigate is enabled.  One of the unified-diff tests now uses
--file-modified-label=comparing: to generate the same output as before,
while the other now uses --navigate to get a delta char label.