Change rules for non-emph style
Fixes #776

Previously, when a paired plus line had no edits, it received
plus-style. With this commit such a line receives
plus-non-emph-style.

There's no change to unpaired lines (still plus-style) and paired
lines with edits (still a mosaic of plus-non-emph-style and plus-emph-style).

(The above statements hold for minus lines also).

Since *-non-emph-style defaults to *-emph-style, this commit does not
result in any change in output for users using the defaults.