Fix weird but correct line handling
This handler function was returning `false`, thus signaling that it
had not handled the line, when it was not the responsibility of any
other handler to handle the line. It was doing this to rely on the
fall-through handlers determining whether to emit the line or skip it.
But this risks another handler handling it and is a violation of the
contract. It is much more appropriate to make the determination in the
handler itself, emit it if appropriate, and signal that it has been
handled.