Center Align numbers right-ish
There is no such thing as "Center Align" with discrete terminal cells. In
some cases a decision has to be made whether to use the left or the right
cell, e.g. when fitting one char into 4 cells: "_X__" or "__X_".

The format!() center/^ default is left, but when padding numbers these
are now aligned to the right if required. Strings remain left-aligned.