Handle a mode change on a renamed file. (#875)
* Handle a mode change on a renamed file.

I changed the diff parsing to cache the mode info from the old/new mode
lines until the parsing bumps into the start of the actual related diff,
finds the diff line for an unrelated diff, or runs out of input.  This
allows the mode info to be output in conjunction with a file event
instead of as a separate heading (that used to have an empty name when
the mode change was for a rename).

The mode info is passed down into the draw routines as a separate
"addendum" string that the draw code can use as it likes. Currently that
means that it appends the addendum string to the non-raw string in
parens. I imagine that future draw routines could decide to put it in a
separate box or some other per-routine method. There is currently a
single function in src/handlers/draw.rs that joins the strings in the
same way for everyone.

A couple examples of how the new code looks:

Δ foo.rs (mode +x)
───────────────────────────────────────────────────

renamed: old-longer-name → shorter-name (mode +x)
───────────────────────────────────────────────────

Would it look better on its own line?

Δ foo.rs
• mode +x
───────────────────────────────────────────────────

renamed: old-longer-name → shorter-name
• mode +x
───────────────────────────────────────────────────

Would it look better appended after a "•" character?

Δ foo.rs • mode +x
───────────────────────────────────────────────────

renamed: old-longer-name → shorter-name • mode +x
───────────────────────────────────────────────────

Should it be a user option? If so, we can do that later.