add .gitattributes rules for rust files
While delta currently removes git's whitespace error markup from diffs
adding the whitespace rules will mean pre-commit hooks that run `git
diff --check` will reject changes with whitespace errors. In order to
get git to print the correct position of the error when there is an
unwanted tab in the indent following a space it is necessary to turn off
`space-before-tab` otherwise the space will be considered to be an error
instead of the tab.

The whitespace rule could perhaps be applied more widely but the
Makefile and example gitcofig have leading tabs so it is restricted them
to just rust files for now.