Misc tab refactoring (#1424)
* Move tabs logic into utils

* Re-use buffer returned by tabs::expand

* Add TabCfg to configure tabs

Use the String from this config for the tab replacement. This avoids
creating a new String for each processed line.

* Avoid unicode segmentation for each line just to remove a prefix

In some code paths no prefix is removed, and in almost all other
cases the prefix is just ascii.

This simplifies a lot of calls.

* Set default tab with to 8

Editors like vim, emacs, nano and most terminal emulators set
this value as the default tab display width.