Wrap --help output and use pager
Set clap option `max_term_width`, so the help output is by default as
wide as the terminal or this max value. Then manually wrap the
`after_long_help()` text (only on demand) to the same width using
`textwrap`. Also use matching ansi codes in this section.

The help output is now paginated if output is to a terminal.

All code paths flow back to main() so an invoked pager is properly
waited for (by the OutputType drop impl).