Do not panic on superimposition mismatch in release builds
Lines which would trigger a panic and early exit with the message
"String mismatch encountered while superimposing style sections"
now try to print the line as well as possible plus a brief error message.

In debug builds and when running from the target/ dir (such as when
core.pager = "target/release/delta" is set) or when the env var
DELTA_NO_WORKAROUNDS is set this mismatch panics as before.