Try to obtain default dark/light state from the terminal
Use the termbg crate to try to read the current terminal background
theme (light or dark) and use it as the default if no --light or --dark
was specified (or light=true in .gitconfig).

This doesn't work in all cases: delta is frequently used with a pipe, in
which case it's not really possible to obtain the background color using
OSC 11. The fallback is the environment variable COLORFGBG which is
supported by most terminals, but it doesn't immediately update when the
system theme changes (it only applies to newly opened terminals). Still,
I think this is better than assuming a dark background.

See https://github.com/dandavison/delta/issues/447 for discussion.