Repair --default-language, and highlight using full filename
Fixed that the "txt" fallback was used instead of --default-language

And when looking for syntax highlighting, keep the filename around
as long as possible.

Using simple custom logic (filename > 4) a Makefile can be highlight with
make syntax, but a cc or ini file does not get treated like a .cc or .ini file.

Currently the underlying highlighting lib syntect and the sublime syntex
definitions can not make this distinction
(http://www.sublimetext.com/docs/syntax.html).