Fixed line number painting (#247)
Fixed the painting of line numbers so that instead of each column or
side always keeping the same colour, which looks odd, it instead uses
the colour appropriate to the line's status.

In other words, if there has been a line removed, instead of showing
the left (minus) column/side as minus style and the right (plus)
column/side as plus style, both will be shown as minus style.