Support external subcommands: rg, git show, git log (etc.)
The possible command line now is:

  delta <delta-args> [SUBCMD <subcmd-args>]

If the entire command line fails to parse because SUBCMD is unknown,
then try (until the next arg fails) parsing <delta-args> only,
and then parse and call SUBCMD with args, its output is piped into
delta.  Other subcommands also take precedence over the diff/git-diff
mode (`delta a b`, where e.g. a=git and b=show), and any diff call gets
converted into an external subcommand first.

Available are:
  delta rg ..       => rg --json .. | delta
  delta a b ..      => git diff --no-index a b .. | delta
  delta git show .. => git <color-on> show .. | delta

and all other git-CMDS, of which
add -p, blame, checkout -p, diff, grep, log -p, reflog -p, and stash show -p
produce a diff.

Because --json is automatically added for `delta rg ..`, it avoids the
parsing ambiguities of and is easier to type than `rg .. | delta`.

The piping is not done by the shell but delta, so the subcommands are
child processes of delta.