Handle quoted file paths in hunk headers
When core.quotepath is true (the default) then non-ASCII chars
in a file name are quoted. These quotes hide the DIFF_PREFIXES and
"a/1" "b/1" remains as such, instead of becoming "1" "1".

This was interpreted by delta as renamed file. Now these quotes are
removed before the DIFF_PREFIXES are searched.